x86/sgx: Rename fallback labels in sgx_init()
It's hard to add new content to this function because it is time
consuming to match fallback and its cause. Rename labels in a way
that the name of error label refers to the site where failure
happened. This way it is easier to keep on track what is going
on.

Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>
[reinette: Placeholder patch - submitted separately upstream
https://lore.kernel.org/linux-sgx/20210928031350.63464-1-jarkko@kernel.org/
]